+++ core/src/index.js
