+++ core/webpack.config.js
